+++ CHANGELOG.md
@@ -62,6 +62,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Use `withSearchErrors` HOC and `buildRecordsManifest` to display an error when the request URL is exceeded. Refs UIIN-2970.
 * Edit in Linked Data Editor when source=MARC. Refs UIIN-2978.
 * Display shared instance immediately after the instance was shared. Refs UIIN-2969.
+* Disable Move to button if no items selected or no more holdings exists within an instance. Refs UIIN-2995.
 
 ## [11.0.4](https://github.com/folio-org/ui-inventory/tree/v11.0.4) (2024-04-30)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.3...v11.0.4)